initial work on extracting loop
